IB/srp: Reduce failover time
The default value for tl_retry_count is changed from 7 into 3.
Hence with a qp->timedout value of 19 this patch reduces the
default transport layer timeout from about 60s to about 25s. The
purpose of this patch is to reduce the time needed for SCSI error
handling significantly and at the same time to avoid activating
the SCSI error handler on an IB path with a regular BER or due to
brief IB network congestion.

Signed-off-by: Bart Van Assche <bvanassche@acm.org>