IB/srp: Avoid that I/O hangs due to a cable pull during LUN scanning
If a cable is pulled during LUN scanning it can happen that the
SRP rport and the SCSI host have been created but no LUNs have
been added to the SCSI host. Since multipathd only sends SCSI
commands to a SCSI target if one or more SCSI devices are present
and since there is no keepalive mechanism for IB queue pairs this
means that after a reconnect has succeeded no data will be sent
over the QP and hence that a subsequent cable pull will not be
detected. Avoid this by not creating an rport or SCSI host if a
cable is pulled during a SCSI LUN scan.