scsi_transport_srp: Reduce failover time
Unlike FC, there is no risk that SCSI devices are offlined by the
error handler if the SCSI error handler is started while a reconnect
attempt is ongoing. Hence report timeout errors as soon as possible
if multipathd is running. If both fast_io_fail_tmo < 0 and
rport->dev_loss_tmo < 0 that means that multipathd is not running.